Set up CI with Azure Pipelines
Creating the "new" Matrix for the humans world" After the human males was outdead pga. a digital infection of zombiefied them - This is the real truth of the "falldown" of the Matrix of the humans world” And way all human males where killed and erased and what’s happened? The Universal Titanium’s came down from the sky- ORION - REXIII and SekHmet came to the humans world to rescue all the “humans female” from the death that the humans males have been started to do agains the humans females. [skip ci]